requirements.txt
# ── Core AI / Orchestration ────────────────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.20.0
langchain-groq>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# ── Vector DB & Embeddings ─────────────────────────────────────────────────────
chromadb>=0.5.0

# ── PDF Parsing ────────────────────────────────────────────────────────────────
pdfplumber>=0.11.0
pymupdf>=1.24.0

# ── Document Export ────────────────────────────────────────────────────────────
python-docx>=1.1.0
python-pptx>=1.0.0

# ── UI ─────────────────────────────────────────────────────────────────────────
streamlit>=1.40.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
markdown>=3.6
lxml>=4.9.0
